+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * New argument `jlmer(fit = TRUE)`. If `FALSE`, returns the unfit model object.
 
+* New function `jl_dict()` to construct a Dictionary object using `list()` syntax.
+
 * Experimental support for `MixedModels.profile()` via `profilelikelihood()` with a `tidy()` method
 
 * Experimental implementation of an interface to setting OptSummary values via `jlmer(optsum = list(...))`.